+++ utils/constants.py
@@ -3,6 +3,7 @@
 
 import os.path
 
+AREA_FILTERS = ["Current view", "All data"]
 ASSET_TYPES = ["bridges", "culverts", "fixeddrainagelevelareas",
                "groundwaterstations", "leveecrosssections", "manholes",
                "measuringstations", "monitoringwells", "outlets", "overflows",
@@ -21,6 +22,11 @@ ASSET_GEOMETRY_TYPES = {"bridges": "Point", "culverts": "LineString",
 
 # The base_url used for getting the data (JSON)
 BASE_URL = "https://demo.lizard.net/api/v2/"
+
+PAYLOAD_BBOX_KEY_TIMESERIES = "geom__within"
+PAYLOAD_BBOX_KEY_OTHER = "in_bbox"
+PRIVATE = "private"
+PUBLIC = "public"
 
 # Add reference to layer_styles directory
 STYLES_ROOT = os.path.join(os.path.dirname(os.path.realpath(__file__)),